requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
supabase
openai
groq
aiohttp
opencv-python
pydantic
python-dotenv
twilio==8.12.0
httpx==0.27.0


